requirements.txt
Flask
Flask-Cors
google-generativeai
chromadb
lxml
gunicorn